Remove the unused register API
